Move Home screen code into separate file
