+++ src/home/Home.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Divider } from 'semantic-ui-react';
+import { Header, Menu } from 'semantic-ui-react';
 import JoinSession from './JoinSession';
 import CreateSession from './CreateSession';
 import Session from './session/Session';
@@ -9,30 +9,34 @@ import {
   createPoll,
   startPoll
 } from '../utils/requests';
+import './Home.css';
 
 class Home extends Component {
   state = {
+    activeTab: 'CREATED',
     session: null,
-    joinError: null,
     joinLoading: false,
-    createError: null,
-    createLoading: false
+    joinError: null,
+    createLoading: false,
+    createError: null
   }
+
+  handleNavbarTabClick = (e, { name }) => this.setState({ activeTab: name })
 
   joinSession = (code) => {
     joinPoll([code])
       .then((session) => {
         this.setState({
           session: session,
-          joinError: null,
-          joinLoading: false
+          joinLoading: false,
+          joinError: null
         });
       })
       .catch((err) => {
         this.setState({
           session: null,
-          joinError: err.toString(),
-          joinLoading: false
+          joinLoading: false,
+          joinError: err.toString()
         });
       });
   }
@@ -58,49 +62,76 @@ class Home extends Component {
       .then((poll) => {
         this.setState({
           session: poll,
-          createError: null,
-          createLoading: false
+          createLoading: false,
+          createError: null
         });
       })
       .catch((err) => {
         this.setState({
-          createError: err.toString(),
-          createLoading: false
+          createLoading: false,
+          createError: err.toString()
         });
       });
   }
 
   render () {
     const {
+      activeTab,
       session,
-      joinError,
-      createError,
       joinLoading,
-      createLoading
+      joinError,
+      createLoading,
+      createError
     } = this.state;
 
-    if (session) {
-      return (
-        <Session
-          session={session}
-          onDisconnect={this.leaveSession}
-        />
-      );
-    }
+    const contentSection = (activeTab == 'CREATED') ? (
+      <div className='created-section'></div>
+    ) : (
+      <div className='joined-section'></div>
+    );
 
     return (
-      <div>
-        <JoinSession
-          error={joinError}
-          loading={joinLoading}
-          onJoin={this.joinSession}
-        />
-        <Divider hidden />
-        <CreateSession
-          error={createError}
-          loading={createLoading}
-          onCreate={this.createSession}
-        />
+      <div className='app'>
+        <div className='app-header'>
+          <div className='header-section'>
+            <Header as='h1' className='header-title'>Polls</Header>
+            <div className='header-content'>
+              <div className='navbar'>
+                <Menu pointing secondary className='navbar-menu'>
+                  <Menu.Item
+                    content='Created'
+                    name='CREATED'
+                    active={activeTab == 'CREATED'}
+                    onClick={this.handleNavbarTabClick}
+                  />
+                  <Menu.Item
+                    content='Joined'
+                    name='JOINED'
+                    active={activeTab == 'JOINED'}
+                    onClick={this.handleNavbarTabClick}
+                  />
+                </Menu>
+              </div>
+              <div className='session-buttons'>
+                <CreateSession
+                  className='create-session-button'
+                  error={createError}
+                  loading={createLoading}
+                  onCreate={this.createSession}
+                />
+                <div className='buttons-divider'></div>
+                <JoinSession
+                  error={joinError}
+                  loading={joinLoading}
+                  onJoin={this.joinSession}
+                />
+              </div>
+            </div>
+          </div>
+        </div>
+        <div className='app-content'>
+          {contentSection}
+        </div>
       </div>
     );
   }
